+++ api/src/crons/index.js
@@ -4,7 +4,7 @@ const computeGoalsInscription = require("./computeGoalsInscription");
 const noticePushMission = require("./noticePushMission");
 const contratRelance = require("./contratRelance");
 const applicationPending = require("./applicationPending");
-const jeVeuxAiderDaily = require("./JeVeuxAiderDaily");
+const jeVeuxAiderDaily = require("./missionsJVA/JeVeuxAiderDaily");
 const loginAttempts = require("./loginAttempts");
 const syncReferentSupport = require("./syncReferentSupport");
 const syncContactSupport = require("./syncContactSupport");
